Moves JSON tests related to a class out of the json package
