+++ wdtk-datamodel/src/test/java/org/wikidata/wdtk/datamodel/implementation/json/JsonSerializerTest.java
@@ -55,7 +55,7 @@ import com.fasterxml.jackson.databind.ObjectReader;
 
 public class JsonSerializerTest {
 
-	static final DatamodelConverter datamodelConverter = new DatamodelConverter(
+	private static final DatamodelConverter datamodelConverter = new DatamodelConverter(
 			new DataObjectFactoryImpl());
 
 	@Test
